+++ src/main/java/teamport/aether/items/itemtool/ItemToolHammerNotch.java
@@ -1,4 +1,4 @@
-package teamport.aether.items;
+package teamport.aether.items.itemtool;
 
 import net.minecraft.core.entity.player.Player;
 import net.minecraft.core.item.ItemStack;
