+++ src/main/java/teamport/aether/items/itemtool/ItemToolHolystone/ItemToolPickaxeHolystone.java
@@ -1,4 +1,4 @@
-package teamport.aether.items.ItemToolHolystone;
+package teamport.aether.items.itemtool.ItemToolHolystone;
 
 import net.minecraft.core.block.Block;
 import net.minecraft.core.block.Blocks;
@@ -7,19 +7,19 @@ import net.minecraft.core.item.ItemStack;
 import net.minecraft.core.item.material.ToolMaterial;
 import net.minecraft.core.util.helper.Side;
 import net.minecraft.core.world.World;
-import teamport.aether.items.ItemToolShovelAether;
+import teamport.aether.items.itemtool.ItemToolPickaxeAether;
 
 import static teamport.aether.items.AetherItems.AMBROSIUM;
 
-public class ItemToolShovelHolystone extends ItemToolShovelAether {
+public class ItemToolPickaxeHolystone extends ItemToolPickaxeAether {
 
-    public ItemToolShovelHolystone(String name, String namespaceId, int id, ToolMaterial enumtoolmaterial) {
+    public ItemToolPickaxeHolystone(String name, String namespaceId, int id, ToolMaterial enumtoolmaterial) {
         super(name, namespaceId, id, enumtoolmaterial);
     }
 
     @Override
-    public boolean onBlockDestroyed(World world, ItemStack itemstack, int i, int x, int y, int z, Side side, Mob
-            mob) {
+
+    public boolean onBlockDestroyed(World world, ItemStack itemstack, int i, int x, int y, int z, Side side, Mob mob) {
         Block<?> block = Blocks.blocksList[i];
         if (block != null && (block.getHardness() > 0.0F || this.isSilkTouch())) {
             itemstack.damageItem(1, mob);
@@ -29,4 +29,5 @@ public class ItemToolShovelHolystone extends ItemToolShovelAether {
         }
         return true;
     }
+
 }
