+++ src/main/java/teamport/aether/items/itemtool/ItemToolPickaxeAether.java
@@ -1,4 +1,4 @@
-package teamport.aether.items;
+package teamport.aether.items.itemtool;
 
 import net.minecraft.core.block.Block;
 import net.minecraft.core.entity.Mob;
