+++ src/main/java/teamport/aether/items/itemtool/ItemToolSwordLightning.java
@@ -1,19 +1,20 @@
-package teamport.aether.items;
+package teamport.aether.items.itemtool;
 
+import net.minecraft.core.entity.EntityLightning;
 import net.minecraft.core.entity.Mob;
 import net.minecraft.core.item.ItemStack;
 import net.minecraft.core.item.material.ToolMaterial;
 import net.minecraft.core.item.tool.ItemToolSword;
 
-public class ItemToolSwordVampire extends ItemToolSword {
+public class ItemToolSwordLightning extends ItemToolSword {
 
-    public ItemToolSwordVampire(String name, String namespaceId, int id, ToolMaterial enumtoolmaterial) {
+    public ItemToolSwordLightning(String name, String namespaceId, int id, ToolMaterial enumtoolmaterial) {
         super(name, namespaceId, id, enumtoolmaterial);
     }
 
     public boolean hitEntity(ItemStack itemstack, Mob target, Mob attacker) {
         if (target instanceof Mob && target.isAlive()) {
-            attacker.heal(8);
+            target.world.entityJoinedWorld(new EntityLightning(target.world, target.x, target.y + 0.5, target.z));
         }
 
         itemstack.damageItem(1, attacker);
