+++ src/main/java/teamport/aether/items/itemtool/ItemToolZanite/ItemToolAxeZanite.java
@@ -1,19 +1,19 @@
-package teamport.aether.items.ItemToolZanite;
+package teamport.aether.items.itemtool.ItemToolZanite;
 
 import net.minecraft.core.block.Block;
 import net.minecraft.core.item.ItemStack;
 import net.minecraft.core.item.material.ToolMaterial;
 import teamport.aether.blocks.AetherBlockTags;
-import teamport.aether.items.ItemToolShovelAether;
+import teamport.aether.items.itemtool.ItemToolAxeAether;
 
-public class ItemToolShovelZanite extends ItemToolShovelAether {
-
-    public ItemToolShovelZanite(String name, String namespaceId, int id, ToolMaterial enumtoolmaterial) {
+public class ItemToolAxeZanite extends ItemToolAxeAether {
+    public ItemToolAxeZanite(String name, String namespaceId, int id, ToolMaterial enumtoolmaterial) {
         super(name, namespaceId, id, enumtoolmaterial);
     }
 
     @Override
     public float getStrVsBlock(ItemStack itemstack, Block<?> block) {
+//        if (!block.hasTag(AetherBlockTags.MINEABLE_BY_AETHER_AXE)) return Items.TOOL_AXE_WOOD.getStrVsBlock(itemstack, block);
         if (!block.hasTag(AetherBlockTags.MINEABLE_BY_AETHER_AXE)) return 1.0F;
         float durability_progress = (float) itemstack.getMetadata() / this.getMaxDamage();
 
@@ -24,3 +24,4 @@ public class ItemToolShovelZanite extends ItemToolShovelAether {
         return (float) (starting_efficiency * (1.0 - durability_progress) + (ending_efficiency * durability_progress));
     }
 }
+
